+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Types of changes:
 
 ## [Unreleased]
 
+## [0.9.0] - 2026-01-16
+
 ### Changed
 
 - Refactor executor to bulk load data into DuckDB memory and run checks from there
@@ -152,7 +154,8 @@ Types of changes:
 - GitHub Actions workflows for CI/CD and releases
 - Use common `defaults` key instead of `global_defaults` and `default_args`
 
-[Unreleased]: https://github.com/ottogroup/koality/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/ottogroup/koality/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/ottogroup/koality/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ottogroup/koality/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ottogroup/koality/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ottogroup/koality/compare/v0.5.0...v0.6.0